Bjó til fleiri pakka/möppur til að skipta klösunum betur niður. Var orðið svo mikið
