+++ app/src/main/java/com/example/sports_app/activities/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.sports_app;
+package com.example.sports_app.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -14,6 +14,7 @@ import android.widget.Toast;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentContainerView;
 
+import com.example.sports_app.R;
 import com.example.sports_app.entities.Thread;
 import com.example.sports_app.entities.User;
 import com.example.sports_app.fragments.SignupFragment;
