+++ app/src/main/java/com/example/sports_app/adapters/EventListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.sports_app;
+package com.example.sports_app.adapters;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -9,13 +9,13 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 
-import com.example.sports_app.entities.Club;
-import com.google.android.material.snackbar.Snackbar;
+import com.example.sports_app.R;
+import com.example.sports_app.entities.Event;
 
 import java.util.ArrayList;
 
-public class ClubListAdapter extends ArrayAdapter<Club> implements View.OnClickListener {
-    private ArrayList<Club> clubList;
+public class EventListAdapter extends ArrayAdapter<Event> implements View.OnClickListener {
+    private ArrayList<Event> eventList;
     Context mContext;
     private int lastPosition = -1;
 
@@ -27,39 +27,39 @@ public class ClubListAdapter extends ArrayAdapter<Club> implements View.OnClickL
 
     @Override
     public View getView(int position, View convertView, ViewGroup container) {
-        Club club = getItem(position);
-        ClubListAdapter.ViewHolder viewHolder;
+        Event event = getItem(position);
+        EventListAdapter.ViewHolder viewHolder;
         final View result;
 
         // Check whether a view object can be reused to improve performance.
         // If not, generate a new view object at this position.
         if (convertView == null) {
-            viewHolder = new ClubListAdapter.ViewHolder();
+            viewHolder = new EventListAdapter.ViewHolder();
             LayoutInflater i = LayoutInflater.from(getContext());
-            convertView = i.inflate(R.layout.clubs_list_row_item, container, false);
-            viewHolder.txtSport = (TextView) convertView.findViewById(R.id.club_sport);
-            viewHolder.txtName = (TextView) convertView.findViewById(R.id.club_name);
+            convertView = i.inflate(R.layout.events_list_row_item, container, false);
+            viewHolder.txtSport = (TextView) convertView.findViewById(R.id.event_sport);
+            viewHolder.txtName = (TextView) convertView.findViewById(R.id.event_name);
 
             result = convertView;
 
             convertView.setTag(viewHolder);
         } else {
-            viewHolder = (ClubListAdapter.ViewHolder) convertView.getTag();
+            viewHolder = (EventListAdapter.ViewHolder) convertView.getTag();
             result = convertView;
         }
 
         lastPosition = position;
 
-        viewHolder.txtSport.setText(club.getmSport());
-        viewHolder.txtName.setText(club.getmClubName());
+        viewHolder.txtSport.setText(event.getSport());
+        viewHolder.txtName.setText(event.getEventName());
 
         viewHolder.txtName.setTag(position);
         return convertView;
     }
 
-    public ClubListAdapter(ArrayList<Club> clubs, @NonNull Context context) {
-        super(context, R.layout.clubs_list_row_item, clubs);
-        this.clubList = clubs;
+    public EventListAdapter(ArrayList<Event> events, @NonNull Context context) {
+        super(context, R.layout.events_list_row_item, events);
+        this.eventList = events;
         this.mContext = context;
     }
 
@@ -67,6 +67,6 @@ public class ClubListAdapter extends ArrayAdapter<Club> implements View.OnClickL
     public void onClick(View v) {
         int position = (Integer) v.getTag();
         Object obj = getItem(position);
-        Club club = (Club) obj;
+        Event event = (Event)obj;
     }
 }
