+++ app/src/main/res/layout/activity_login.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".LoginActivity">
+    tools:context=".activities.LoginActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/sign_up_fragment_container_view"
