+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:gravity="top|center"
     android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context=".activities.MainActivity">
 
     <LinearLayout
         android:layout_height="wrap_content"
