+++ app/src/main/res/menu/menu_actionview.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_menu"
-    tools:context=".MainActivity">
+    tools:context=".activities.MainActivity">
 
     <item android:id="@+id/menu_login"
         android:title="@string/menu_login"/>
